+++ content/lib/pyswitch/hardware/devices/pa_midicaptain_10.py
@@ -51,7 +51,7 @@ PA_MIDICAPTAIN_10_SWITCH_A      = { "model": _AdafruitSwitch(_board.GP9),  "pixe
 PA_MIDICAPTAIN_10_SWITCH_B      = { "model": _AdafruitSwitch(_board.GP10), "pixels": (18, 20, 19), "name": "B", "strobeOrder": 8 }
 PA_MIDICAPTAIN_10_SWITCH_C      = { "model": _AdafruitSwitch(_board.GP11), "pixels": (21, 23, 22), "name": "C", "strobeOrder": 7 }
 PA_MIDICAPTAIN_10_SWITCH_D      = { "model": _AdafruitSwitch(_board.GP18), "pixels": (24, 26, 25), "name": "D", "strobeOrder": 6 }
-PA_MIDICAPTAIN_10_SWITCH_DOWN   = { "model": _AdafruitSwitch(_board.GP19), "pixels": (27, 29, 28), "name": "Dn", "strobeOrder": 5 }
+PA_MIDICAPTAIN_10_SWITCH_DOWN   = { "model": _AdafruitSwitch(_board.GP19), "pixels": (27, 29, 28), "name": "Down", "strobeOrder": 5 }
 
 PA_MIDICAPTAIN_10_EXP_PEDAL_1   = { "model": _AdafruitPotentiometer(_board.GP27), "name": "Exp1" }
 PA_MIDICAPTAIN_10_EXP_PEDAL_2   = { "model": _AdafruitPotentiometer(_board.GP28), "name": "Exp2" }