+++ test/pyswitch/test_input_controller_switch_hold.py
@@ -100,7 +100,7 @@ class TestSwitchControllerHold(unittest.TestCase):
         self.assertEqual(action_3.num_release_calls, 0)
 
         hold_period.exceed_next_time = True
-        
+        fs.process()
         switch.shall_be_pushed = False
         fs.process()
 
@@ -162,6 +162,344 @@ class TestSwitchControllerHold(unittest.TestCase):
 
         self.assertEqual(action_3.num_push_calls, 2)
         self.assertEqual(action_3.num_release_calls, 2)
+
+        switch.shall_be_pushed = False
+        fs.process()
+
+        # Repeated long press (must not work here)
+        switch.shall_be_pushed = True
+        fs.process()
+        hold_period.exceed_next_time = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 3)
+        self.assertEqual(action_3.num_release_calls, 3)
+
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 3)
+        self.assertEqual(action_3.num_release_calls, 3)
+
+        hold_period.exceed_next_time = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 3)
+        self.assertEqual(action_3.num_release_calls, 3)
+
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 3)
+        self.assertEqual(action_3.num_release_calls, 3)
+
+        hold_period.exceed_next_time = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 3)
+        self.assertEqual(action_3.num_release_calls, 3)
+
+        # Let off again
+        hold_period.exceed_next_time = False
+        switch.shall_be_pushed = False
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 3)
+        self.assertEqual(action_3.num_release_calls, 3)
+
+        # Short press again
+        switch.shall_be_pushed = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 3)
+        self.assertEqual(action_3.num_release_calls, 3)
+
+        switch.shall_be_pushed = False
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 3)
+        self.assertEqual(action_1.num_release_calls, 3)
+        
+        self.assertEqual(action_2.num_push_calls, 3)
+        self.assertEqual(action_2.num_release_calls, 3)
+
+        self.assertEqual(action_3.num_push_calls, 3)
+        self.assertEqual(action_3.num_release_calls, 3)
+
+
+    def test_hold_repeat(self):
+        hold_period = MockPeriodCounter()
+        switch = MockSwitch()
+
+        action_1 = MockAction()
+        action_2 = MockAction()
+        action_3 = MockAction()
+    
+        appl = MockController2()
+
+        fs = SwitchController(
+            appl,
+            {
+                "assignment": {
+                    "model": switch
+                },
+                "actions": [
+                    action_1,
+                    action_2
+                ],
+                "actionsHold": [
+                    action_3
+                ],
+                "holdRepeat": True
+            },
+            hold_period
+        )
+
+        self.assertEqual(fs.actions, [action_1, action_2, action_3])
+
+        # Short press
+        switch.shall_be_pushed = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 0)
+        self.assertEqual(action_1.num_release_calls, 0)
+
+        self.assertEqual(action_2.num_push_calls, 0)
+        self.assertEqual(action_2.num_release_calls, 0)
+
+        self.assertEqual(action_3.num_push_calls, 0)
+        self.assertEqual(action_3.num_release_calls, 0)
+
+        switch.shall_be_pushed = False
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 1)
+        self.assertEqual(action_1.num_release_calls, 1)
+        
+        self.assertEqual(action_2.num_push_calls, 1)
+        self.assertEqual(action_2.num_release_calls, 1)
+
+        self.assertEqual(action_3.num_push_calls, 0)
+        self.assertEqual(action_3.num_release_calls, 0)
+        
+        # Long press
+        switch.shall_be_pushed = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 1)
+        self.assertEqual(action_1.num_release_calls, 1)
+        
+        self.assertEqual(action_2.num_push_calls, 1)
+        self.assertEqual(action_2.num_release_calls, 1)
+
+        self.assertEqual(action_3.num_push_calls, 0)
+        self.assertEqual(action_3.num_release_calls, 0)
+
+        hold_period.exceed_next_time = True
+        fs.process()
+        switch.shall_be_pushed = False
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 1)
+        self.assertEqual(action_1.num_release_calls, 1)
+        
+        self.assertEqual(action_2.num_push_calls, 1)
+        self.assertEqual(action_2.num_release_calls, 1)
+
+        self.assertEqual(action_3.num_push_calls, 1)
+        self.assertEqual(action_3.num_release_calls, 1)
+
+        # Short press again
+        switch.shall_be_pushed = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 1)
+        self.assertEqual(action_1.num_release_calls, 1)
+        
+        self.assertEqual(action_2.num_push_calls, 1)
+        self.assertEqual(action_2.num_release_calls, 1)
+
+        self.assertEqual(action_3.num_push_calls, 1)
+        self.assertEqual(action_3.num_release_calls, 1)
+
+        switch.shall_be_pushed = False
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 1)
+        self.assertEqual(action_3.num_release_calls, 1)
+
+        # Long press with running out of time before release
+        switch.shall_be_pushed = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 1)
+        self.assertEqual(action_3.num_release_calls, 1)
+
+        hold_period.exceed_next_time = True        
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 2)
+        self.assertEqual(action_3.num_release_calls, 2)
+
+        switch.shall_be_pushed = False
+        fs.process()
+
+        # Repeated long press
+        switch.shall_be_pushed = True
+        fs.process()
+        hold_period.exceed_next_time = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 3)
+        self.assertEqual(action_3.num_release_calls, 3)
+
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 3)
+        self.assertEqual(action_3.num_release_calls, 3)
+
+        hold_period.exceed_next_time = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 4)
+        self.assertEqual(action_3.num_release_calls, 4)
+
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 4)
+        self.assertEqual(action_3.num_release_calls, 4)
+
+        hold_period.exceed_next_time = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 5)
+        self.assertEqual(action_3.num_release_calls, 5)
+
+        # Let off again
+        switch.shall_be_pushed = False
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 5)
+        self.assertEqual(action_3.num_release_calls, 5)
+
+        # Short press again
+        switch.shall_be_pushed = True
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 2)
+        self.assertEqual(action_1.num_release_calls, 2)
+        
+        self.assertEqual(action_2.num_push_calls, 2)
+        self.assertEqual(action_2.num_release_calls, 2)
+
+        self.assertEqual(action_3.num_push_calls, 5)
+        self.assertEqual(action_3.num_release_calls, 5)
+
+        switch.shall_be_pushed = False
+        fs.process()
+
+        self.assertEqual(action_1.num_push_calls, 3)
+        self.assertEqual(action_1.num_release_calls, 3)
+        
+        self.assertEqual(action_2.num_push_calls, 3)
+        self.assertEqual(action_2.num_release_calls, 3)
+
+        self.assertEqual(action_3.num_push_calls, 5)
+        self.assertEqual(action_3.num_release_calls, 5)
 
 
     def test_no_hold(self):